LayoutTests/inspector/elements/undo-dom-edits-expected.txt

Tests that DOM modifications done in the Elements panel are undoable.


Running: testSetUp

Running: testRemove
Initial:
- <div id="testRemove">
      <div id="node-to-remove"></div>
  </div>
Post-action:
- <div id="testRemove">
  </div>
Post-undo (initial):
- <div id="testRemove">
      <div id="node-to-remove"></div>
  </div>

Running: testSetNodeName
Initial:
- <div id="testSetNodeName">
      <div id="node-to-set-name"></div>
  </div>
Post-action:
- <div id="testSetNodeName">
      <span id="node-to-set-name"></span>
  </div>
Post-undo (initial):
- <div id="testSetNodeName">
      <div id="node-to-set-name"></div>
  </div>

Running: testSetNodeValue
Initial:
- <div id="testSetNodeValue">
      <div id="node-to-set-value">Text</div>
  </div>
Post-action:
- <div id="testSetNodeValue">
      <div id="node-to-set-value">New Text</div>
  </div>
Post-undo (initial):
- <div id="testSetNodeValue">
      <div id="node-to-set-value">Text</div>
  </div>

Running: testEditAsHTML
Initial:
- <div id="testEditAsHTML">
    - <div id="node-to-edit-as-html">
          <span id="span">Text</span>
      </div>
  </div>
Post-action:
- <div id="testEditAsHTML">
    - <div id="node-to-edit-as-html">
          <div id="span2">Text2</div>
      </div>
  </div>
Post-undo (initial):
- <div id="testEditAsHTML">
    - <div id="node-to-edit-as-html">
          <span id="span">Text</span>
      </div>
  </div>
